\name{venn-internal}

\alias{.onAttach}
\alias{checkZone}
\alias{openPlot}
\alias{plotRules}


\alias{borders}
\alias{ib}
\alias{ints}
\alias{icoords}

\title{Internal Functions}

\description{
  The above functions are internal in the venn package which are not designed to 
  be called directly by the user. 
}

\keyword{internal}