\name{venn-package}

\alias{venn-package}


\docType{package}

\title{
Draw Venn Diagrams
}

\description{
Draws and displays Venn diagrams up to 7 sets, and any boolean union of set intersections.
}

\details{
    \tabular{ll}{
    Package: \tab venn\cr
    Type:    \tab Package\cr
    Version: \tab 1.10\cr
    Date:    \tab 2021-03-15\cr
    License: \tab GPL (>= 2)\cr
    }
}

\author{

\bold{Authors}:\cr
Adrian Dusa\cr
Department of Sociology\cr
University of Bucharest\cr
\email{dusa.adrian@unibuc.ro}

\bold{Maintainer}:\cr
Adrian Dusa
}


\keyword{package}
