\name{venn internal functions}

\alias{checkZone}
\alias{openPlot}
\alias{translate2}
\alias{plotRules}
\alias{getBorders}
\alias{getIB}
\alias{getICoords}
\alias{getSCoords}
\alias{getInts}
\alias{getSets}


\title{Internal Functions}

\description{
  The above functions are internal in the venn package,
  which are not designed to be called directly by the user.
}

\keyword{internal}
