\name{About the venn package}

\alias{venn-package}

\docType{package}

\title{
Draw Venn Diagrams
}

\description{
A close to zero dependency package to draw and display Venn diagrams, and any
boolean union of set intersections.
}

\details{
    \tabular{ll}{
    Package: \tab venn\cr
    Type:    \tab Package\cr
    Version: \tab 1.11-6\cr
    Date:    \tab 2024-01-07\cr
    License: \tab GPL (>= 2)\cr
    }
}

\author{

\bold{Authors}:\cr
Adrian Dusa\cr
Department of Sociology\cr
University of Bucharest\cr
\email{dusa.adrian@unibuc.ro}

\bold{Maintainer}:\cr
Adrian Dusa
}


\keyword{package}
