\name{venn internal functions}

\alias{.onAttach}
\alias{checkZone}
\alias{openPlot}
\alias{translate2}
\alias{plotRules}


\title{Internal Functions}

\description{
  The above functions are internal in the venn package,
  which are not designed to be called directly by the user. 
}

\keyword{internal}
